Figuring out a lot

+++ src/SonOfPicasso.Core/Services/ImageContainerImageManagementService.cs
@@ -2,9 +2,12 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Reactive;
+using System.Reactive.Disposables;
 using System.Reactive.Linq;
+using System.Reactive.Subjects;
 using DynamicData;
 using Serilog;
+using SonOfPicasso.Core.Extensions;
 using SonOfPicasso.Core.Interfaces;
 using SonOfPicasso.Core.Model;
 using SonOfPicasso.Core.Scheduling;
@@ -14,15 +17,15 @@ namespace SonOfPicasso.Core.Services
 {
     public class ImageContainerImageManagementService : IImageContainerManagementService
     {
+        private readonly Subject<IObservable<FileSystemEventArgs>> _currentFileWatcherSubject;
+        private readonly CompositeDisposable _disposables;
         private readonly IFolderRulesManagementService _folderRulesManagementService;
         private readonly IFolderWatcherService _folderWatcherService;
-
         private readonly SourceCache<IImageContainer, string> _imageContainerCache;
         private readonly IImageContainerOperationService _imageContainerOperationService;
         private readonly IObservableCache<ImageRef, string> _imageRefCache;
         private readonly ILogger _logger;
         private readonly ISchedulerProvider _schedulerProvider;
-        private IDisposable _folderManagementDisposable;
 
         public ImageContainerImageManagementService(
             IImageContainerOperationService imageContainerOperationService,
@@ -42,6 +45,13 @@ namespace SonOfPicasso.Core.Services
                 .ObserveOn(schedulerProvider.TaskPool)
                 .TransformMany(container => container.ImageRefs, imageRef => imageRef.Id)
                 .AsObservableCache();
+
+            _disposables = new CompositeDisposable();
+
+            _currentFileWatcherSubject = new Subject<IObservable<FileSystemEventArgs>>();
+            _currentFileWatcherSubject.Switch()
+                .Subscribe(HandlerFolderWatcherEvent)
+                .DisposeWith(_disposables);
         }
 
         public IConnectableCache<IImageContainer, string> ImageContainerCache => _imageContainerCache;
@@ -50,9 +60,16 @@ namespace SonOfPicasso.Core.Services
 
         public void Dispose()
         {
-            _folderManagementDisposable?.Dispose();
             _imageContainerCache?.Dispose();
             _imageRefCache?.Dispose();
+            _currentFileWatcherSubject?.Dispose();
+            _disposables?.Dispose();
+        }
+
+        public IObservable<Unit> ResetFolderManagementRules(IEnumerable<IFolderRuleInput> folderRules)
+        {
+            return _folderRulesManagementService.ResetFolderManagementRules(folderRules)
+                .Do(unit => StartWatcher());
         }
 
         public IObservable<Unit> Start()
@@ -66,8 +83,7 @@ namespace SonOfPicasso.Core.Services
 
         public void Stop()
         {
-            _folderManagementDisposable?.Dispose();
-            _folderManagementDisposable = null;
+            _currentFileWatcherSubject.OnNext(Observable.Never<FileSystemEventArgs>());
         }
 
         public IObservable<IImageContainer> ScanFolder(string path)
@@ -128,39 +144,40 @@ namespace SonOfPicasso.Core.Services
 
         private void StartWatcher()
         {
-            _folderManagementDisposable = _folderRulesManagementService.GetFolderManagementRules()
-                .SelectMany(list => _folderWatcherService.WatchFolders(list, Constants.ImageExtensions))
-                .Subscribe(HandlerFolderWatcherEvent);
+            var observable = _folderRulesManagementService.GetFolderManagementRules()
+                .SelectMany(list => _folderWatcherService.WatchFolders(list, Constants.ImageExtensions));
+
+            _currentFileWatcherSubject.OnNext(observable);
         }
 
         private void HandlerFolderWatcherEvent(FileSystemEventArgs args)
         {
-            _logger.Debug("HandlerFolderWatcherEvent {ChangeType} {Name}", args.ChangeType, args.Name);
-
-            switch (args.ChangeType)
+            if (args.ChangeType != WatcherChangeTypes.Created)
             {
-                case WatcherChangeTypes.Created:
-                    break;
+                _logger.Debug("HandlerFolderWatcherEvent {ChangeType} {Name}", args.ChangeType, args.Name);
 
-                case WatcherChangeTypes.Changed:
-                    _imageContainerCache.PopulateFrom(
-                        _imageContainerOperationService.AddOrUpdateImage(args.FullPath));
-                    break;
+                switch (args.ChangeType)
+                {
+                    case WatcherChangeTypes.Changed:
+                        _imageContainerCache.PopulateFrom(
+                            _imageContainerOperationService.AddOrUpdateImage(args.FullPath));
+                        break;
 
-                case WatcherChangeTypes.Deleted:
-                    _imageContainerCache.PopulateFrom(
-                        _imageContainerOperationService.DeleteImage(args.FullPath));
-                    break;
+                    case WatcherChangeTypes.Deleted:
+                        _imageContainerCache.PopulateFrom(
+                            _imageContainerOperationService.DeleteImage(args.FullPath));
+                        break;
 
-                case WatcherChangeTypes.Renamed:
-                    var renamedEventArgs = (RenamedEventArgs) args;
-                    _imageContainerCache.PopulateFrom(
-                        _imageContainerOperationService.RenameImage(renamedEventArgs.OldFullPath,
-                            renamedEventArgs.FullPath));
-                    break;
+                    case WatcherChangeTypes.Renamed:
+                        var renamedEventArgs = (RenamedEventArgs) args;
+                        _imageContainerCache.PopulateFrom(
+                            _imageContainerOperationService.RenameImage(renamedEventArgs.OldFullPath,
+                                renamedEventArgs.FullPath));
+                        break;
 
-                default:
-                    throw new ArgumentOutOfRangeException();
+                    default:
+                        throw new ArgumentOutOfRangeException();
+                }
             }
         }
     }
